Chore : Board 관련 중간 커밋
main 수복 위해 임시 commit

+++ spring_team_prj/src/main/java/com/sweetHome/vo/BoardImagesVO.java
@@ -11,6 +11,6 @@ public class BoardImagesVO {
 	private int boardSeq;
 	private String oname;
 	private String sname;
-	private int fsize;
+	private long fsize;
 	private String fpath;
 }
