+++ spring_team_prj/src/main/java/com/sweetHome/vo/CommunityImagesVO.java
@@ -10,7 +10,7 @@ public class CommunityImagesVO {
 	private int communityImageSeq;
 	private String oname;
 	private String sname;
-	private int fsize;
+	private long fsize;
 	private String fpath;
 	private int communitySeq;
 }
